Flytter api-kall for ferdigstilling inn i et hook og tar hooket i bruk i Brev, OmgjørVedtak og BrevFaneUtenBrev
